requirements.txt
streamlit
langchain
faiss-cpu
pymupdf
requests
transformers
sentence-transformers
python-dotenv
langchain-community
